@@ -1,14 +1,25 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter } from 'react-router-dom';
+// import {
+//   QueryClient,
+//   QueryClientProvider,
+//   // useQuery,
+// } from 'react-query';
+// const queryClient = new QueryClient();
 
-ReactDOM.render(
-  <React.Fragment>
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    {/* <QueryClientProvider client={queryClient}> */}
+    <BrowserRouter>
     <App />
-  </React.Fragment>,
-  document.getElementById('root')
+    </BrowserRouter>
+    {/* </QueryClientProvider> */}
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
